+++ crates/voicevox_core/src/engine/talk/audio_query.rs
@@ -1,8 +1,23 @@
+mod validated;
+
 use serde::{Deserialize, Serialize};
+
+pub(crate) use self::validated::{
+    LengthedPhoneme, ValidatedAccentPhrase, ValidatedAudioQuery, ValidatedMora,
+};
+
+pub use self::validated::Validate;
 
 /* 各フィールドのjsonフィールド名はsnake_caseとする*/
 
 /// モーラ（子音＋母音）ごとの情報。
+///
+/// # Validation
+///
+/// この構造体の状態によっては、`Synthesizer`の各メソッドは[`ErrorKind::InvalidQuery`]を表わすエラーを返す。詳細は[`validate`メソッド]にて。
+///
+/// [`ErrorKind::InvalidQuery`]: crate::ErrorKind::InvalidQuery
+/// [`validate`メソッド]: Self::validate
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
 #[non_exhaustive]
 pub struct Mora {
@@ -21,6 +36,13 @@ pub struct Mora {
 }
 
 /// AccentPhrase (アクセント句ごとの情報)。
+///
+/// # Validation
+///
+/// この構造体の状態によっては、`Synthesizer`の各メソッドは[`ErrorKind::InvalidQuery`]を表わすエラーを返す。詳細は[`validate`メソッド]にて。
+///
+/// [`ErrorKind::InvalidQuery`]: crate::ErrorKind::InvalidQuery
+/// [`validate`メソッド]: Self::validate
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
 #[non_exhaustive]
 pub struct AccentPhrase {
@@ -52,6 +74,13 @@ impl AccentPhrase {
 /// VOICEVOX ENGINEと同じスキーマになっている。ただし今後の破壊的変更にて変わる可能性がある。[データのシリアライゼーション]を参照。
 ///
 /// [データのシリアライゼーション]: https://github.com/VOICEVOX/voicevox_core/blob/main/docs/guide/user/serialization.md
+///
+/// # Validation
+///
+/// この構造体の状態によっては、`Synthesizer`の各メソッドは[`ErrorKind::InvalidQuery`]を表わすエラーを返す。詳細は[`validate`メソッド]にて。
+///
+/// [`ErrorKind::InvalidQuery`]: crate::ErrorKind::InvalidQuery
+/// [`validate`メソッド]: Self::validate
 #[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
 #[non_exhaustive]
 pub struct AudioQuery {