+++ crates/voicevox_core_python_api/src/convert.rs
@@ -23,10 +23,11 @@ use voicevox_core::{
 
 use crate::{
     _ReservedFields, AnalyzeTextError, GetSupportedDevicesError, GpuSupportError,
-    InitInferenceRuntimeError, InvalidModelDataError, InvalidModelFormatError, InvalidWordError,
-    LoadUserDictError, ModelAlreadyLoadedError, ModelNotFoundError, NotLoadedOpenjtalkDictError,
-    OpenZipFileError, ParseKanaError, ReadZipEntryError, RunModelError, SaveUserDictError,
-    StyleAlreadyLoadedError, StyleNotFoundError, UseUserDictError, WordNotFoundError,
+    InitInferenceRuntimeError, InvalidModelDataError, InvalidModelFormatError, InvalidQueryError,
+    InvalidWordError, LoadUserDictError, ModelAlreadyLoadedError, ModelNotFoundError,
+    NotLoadedOpenjtalkDictError, OpenZipFileError, ParseKanaError, ReadZipEntryError,
+    RunModelError, SaveUserDictError, StyleAlreadyLoadedError, StyleNotFoundError,
+    UseUserDictError, WordNotFoundError,
 };
 
 pub(crate) fn from_acceleration_mode(ob: &Bound<'_, PyAny>) -> PyResult<AccelerationMode> {
@@ -247,7 +248,9 @@ impl<'py> IntoPyObject<'py> for ToPyUuid {
     }
 }
 
-fn from_dataclass_via_serde<T: DeserializeOwned>(instance: &Bound<'_, PyAny>) -> PyResult<T> {
+pub(crate) fn from_dataclass_via_serde<T: DeserializeOwned>(
+    instance: &Bound<'_, PyAny>,
+) -> PyResult<T> {
     let fields = dataclasses_asdict(instance)?;
     serde_pyobject::from_pyobject(fields).map_err(Into::into)
 }
@@ -383,6 +386,7 @@ pub(crate) impl<T> voicevox_core::Result<T> {
                 ErrorKind::WordNotFound => WordNotFoundError::new_err(msg),
                 ErrorKind::UseUserDict => UseUserDictError::new_err(msg),
                 ErrorKind::InvalidWord => InvalidWordError::new_err(msg),
+                ErrorKind::InvalidQuery => InvalidQueryError::new_err(msg),
                 ErrorKind::__NonExhaustive => unreachable!(),
             };